Update Scalar class with new functionaliity; Remove Variable class
